+++ module/canvas/torgeternityToken.js
@@ -1,3 +1,5 @@
+import TorgeternityActor from '../documents/actor/torgeternityActor.js';
+
 export default class TorgEternityToken extends foundry.canvas.placeables.Token {
 
   static #originToken;
@@ -10,7 +12,7 @@ export default class TorgEternityToken extends foundry.canvas.placeables.Token {
     if (!game.settings.get('torgeternity', 'hoverDistance') || !this.visible) return result;
     if (!canvas.tokens.active || game.activeTool !== "select") return result;
 
-    const originActor = game.user.character;
+    const originActor = TorgeternityActor.getControlledActor();
     if (!originActor) return result;
     const originToken = canvas.scene.tokens.find(t => t.actor === originActor);
     if (!originToken || originToken.object === this) return;